requirements.txt
Pillow==6.2.1
pytest==5.3.2